requirements.txt
tk
pyserial
RPi.GPIO; sys_platform == 'linux'
